Changes_made_to_Quants
Made changes in Gemma_Test and Gemma_Test_2 files. The Quantized models are on my HF repo. I still have to try them out and see if they are efficient. Regarding the Quant_test file, I am having some issues with the libraries, so if anyone can try them, and let me know if they work, that would be great.